% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpipollV2-package.R
\docType{package}
\name{SpipollV2-package}
\alias{SpipollV2}
\alias{SpipollV2-package}
\title{SpipollV2: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/charles-thevenin/SpipollV2}
  \item Report bugs at \url{https://github.com/charles-thevenin/SpipollV2/issues}
}

}
\author{
\strong{Maintainer}: Charles Thévenin \email{thevenin.charles@gmail.com} (\href{https://orcid.org/0000-0002-5520-476X}{ORCID}) [copyright holder]

}
\keyword{internal}
